refactor: Participant 클래스가 cardList를 소유하도록 구현
